+++ src/app/toolbar/toolbar.component.ts
@@ -1,11 +1,10 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input, inject } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatTooltipModule } from '@angular/material/tooltip';
-import { Modeler } from '@denysvuika/ng-bpmn';
-import { saveAs } from 'file-saver';
+import { Modeler, ModelingService } from '@denysvuika/ng-bpmn';
 
 @Component({
   standalone: true,
@@ -14,25 +13,19 @@ import { saveAs } from 'file-saver';
   templateUrl: './toolbar.component.html'
 })
 export class AppToolbarComponent {
+  private modelingService = inject(ModelingService);
+
   @Input({ required: true }) modeler?: Modeler;
 
   async saveXML(modeler?: Modeler) {
     if (modeler) {
-      const content = await modeler.saveXML();
-      if (content) {
-        const blob = new Blob([content]);
-        saveAs(blob, 'diagram.xml');
-      }
+      this.modelingService.downloadXML(modeler);
     }
   }
 
   async saveSVG(modeler?: Modeler) {
     if (modeler) {
-      const content = await modeler.saveSVG();
-      if (content) {
-        const blob = new Blob([content], { type: 'image/svg+xml' });
-        saveAs(blob, 'diagram.svg');
-      }
+      this.modelingService.downloadSVG(modeler);
     }
   }
 }