+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- new `ModelingService` to handle frequently used modeling actions (i.e. export to SVG/XML, etc.)
+
+### Changed
+
 ## [0.0.3] - 2023-05-13
 
 ### Added